ssl: only set "sslrootcert" once
Fixes #570, though I have not been able to write a test for this.

Remove methods on `values`. These methods are not a useful abstraction
and only serve to obscure that the underlying structure is a map. Their
presence is particularly inconvenient in that they promote confusing
empty values with the lack of value.

While I'm here, also improve conformance with libpq's treatment of
blank sslcert and sslkey option values. See the test changes for
specific changes.